+++ Assets/OxGKit/Utilities/Scripts/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+## [0.0.8-preview] - 2023-10-16
+- Added Remove method for Cacher.
+- Added Remove method for Requester (can remove specific url from cache).
+
 ## [0.0.7-preview] - 2023-09-11
 - Added InterTimer.
 - Added IntervalSetter (global factory).